Migrate DishPicker to MobX
