+++ src/frontend/js/stores/Dishes.js
@@ -1,0 +1,7 @@
+import Collection from './Collection';
+
+export default class Dishes extends Collection {
+  get url() {
+    return 'dishes';
+  }
+}
