+++ src/frontend/js/stores/index.js
@@ -1,12 +1,14 @@
 import { autorun } from 'mobx';
 
 import DateRange from './DateRange';
+import Dishes from './Dishes';
 import Meals from './Meals';
 import People from './People';
 import PlannedMeals from './PlannedMeals';
 import Security from './Security';
 
 const dateRange = new DateRange();
+const dishes = new Dishes();
 const people = new People();
 const security = new Security();
 
@@ -19,10 +21,13 @@ autorun(() => {
     meals.fetch();
     plannedMeals.fetch();
   }
-})
+});
+
+dishes.fetch();
 
 export default {
   dateRange,
+  dishes,
   meals,
   people,
   plannedMeals,
